Rename module oauth2-provider-scopes-api to oauth2-apps-api
