+++ modules/apps/oauth2-provider/oauth2-apps-api/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay OAuth2 Apps API
+Bundle-SymbolicName: com.liferay.oauth2.apps.api
+Bundle-Version: 1.0.0
+Export-Package:\
+	com.liferay.oauth2.apps.api.*
