+++ modules/apps/oauth2-provider/oauth2-apps-api/build.gradle
