+++ modules/apps/oauth2-provider/oauth2-apps-api/src/main/java/com/liferay/oauth2/apps/api/RequiresScope.java
@@ -1,4 +1,4 @@
-package com.liferay.oauth2.provider.scopes.api;
+package com.liferay.oauth2.apps.api;
 
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
