+++ modules/apps/oauth2-provider/oauth2-provider-rest/build.gradle
@@ -21,8 +21,8 @@ dependencies {
 	compileOnly group: "org.osgi", name: "org.osgi.service.component", version: "1.3.0"
 	compileOnly group: "org.osgi", name: "org.osgi.service.http.whiteboard", version: "1.0.0"
 
+	compileOnly project(":apps:oauth2-provider:oauth2-apps-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-api")
-	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-liferay-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-spi")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-rest-spi")
