+++ modules/apps/oauth2-provider/oauth2-provider-rest/src/main/java/com/liferay/oauth2/provider/rest/requestscopechecker/AnnotationRequestScopeChecker.java
@@ -14,9 +14,9 @@
 
 package com.liferay.oauth2.provider.rest.requestscopechecker;
 
+import com.liferay.oauth2.apps.api.RequiresScope;
+import com.liferay.oauth2.apps.api.ScopeChecker;
 import com.liferay.oauth2.provider.rest.scopechecker.spi.RequestScopeChecker;
-import com.liferay.oauth2.provider.scopes.api.RequiresScope;
-import com.liferay.oauth2.provider.scopes.api.ScopeChecker;
 import com.liferay.portal.kernel.util.MapUtil;
 import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
