+++ modules/apps/oauth2-provider/oauth2-provider-rest/src/test/java/com/liferay/oauth2/provider/scopes/impl/AnnotationRequestScopeCheckerTest.java
@@ -15,7 +15,7 @@
 package com.liferay.oauth2.provider.scopes.impl;
 
 import com.liferay.oauth2.provider.rest.requestscopechecker.AnnotationRequestScopeChecker;
-import com.liferay.oauth2.provider.scopes.api.RequiresScope;
+import com.liferay.oauth2.apps.api.RequiresScope;
 
 import org.junit.Before;
 import org.junit.Test;
