--- modules/apps/oauth2-provider/oauth2-provider-scopes-api/bnd.bnd
@@ -1,5 +1,0 @@
-Bundle-Name: Liferay OAuth2 Provider Scopes API
-Bundle-SymbolicName: com.liferay.oauth2.provider.scopes.api
-Bundle-Version: 1.0.0
-Export-Package:\
-	com.liferay.oauth2.provider.scopes.api.*
