+++ modules/apps/oauth2-provider/oauth2-provider-scopes-impl/src/main/java/com/liferay/oauth2/provider/scopes/impl/feature/ScopeAnnotationFinder.java
@@ -14,12 +14,13 @@
 
 package com.liferay.oauth2.provider.scopes.impl.feature;
 
-import com.liferay.oauth2.provider.scopes.api.RequiresScope;
 import org.apache.cxf.Bus;
 import org.apache.cxf.jaxrs.model.ClassResourceInfo;
 import org.apache.cxf.jaxrs.model.MethodDispatcher;
 import org.apache.cxf.jaxrs.model.OperationResourceInfo;
 import org.apache.cxf.jaxrs.utils.ResourceUtils;
+
+import com.liferay.oauth2.apps.api.RequiresScope;
 
 import java.lang.reflect.Method;
 import java.util.Arrays;
