+++ modules/apps/oauth2-provider/oauth2-provider-scopes-liferay-api/build.gradle
@@ -8,7 +8,7 @@ dependencies {
 	compileOnly group: "com.liferay", name: "com.liferay.osgi.service.tracker.collections", version: "2.0.0"
 	compileOnly group: "javax.ws.rs", name: "javax.ws.rs-api", version: "2.0.1"
 	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
+	compileOnly project(":apps:oauth2-provider:oauth2-apps-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-api")
-	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-api")
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-spi")
 }
