+++ modules/apps/oauth2-provider/oauth2-provider-scopes-sample2/src/main/java/com/liferay/oauth2/provider/sample2/oauth/Test.java
@@ -18,8 +18,8 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonSerializer;
+import com.liferay.oauth2.apps.api.RequiresScope;
 import com.liferay.oauth2.provider.model.LiferayOAuth2Scope;
-import com.liferay.oauth2.provider.scopes.api.RequiresScope;
 
 import java.util.Collection;
 import java.util.Set;
