+++ modules/apps/oauth2-provider/oauth2-provider-scopes-spi/build.gradle
@@ -7,5 +7,5 @@ dependencies {
 	compileOnly group: "com.liferay.portal", name: "com.liferay.portal.kernel", version: "2.6.0"
 	compileOnly group: "org.osgi", name: "org.osgi.core", version: "5.0.0"
 	compileOnly project(":apps:oauth2-provider:oauth2-provider-api")
-	compileOnly project(":apps:oauth2-provider:oauth2-provider-scopes-api")
+	compileOnly project(":apps:oauth2-provider:oauth2-apps-api")
 }
