+++ modules/apps/oauth2-provider/oauth2-provider-scopes-spi/src/main/java/com/liferay/oauth2/provider/scopes/spi/ScopeFinder.java
@@ -30,9 +30,9 @@ public interface ScopeFinder {
 	 * Implementations <b>SHOULD</b> test all their available scopes against
 	 * the matcher to allow for different matching strategies to succeed.
 	 * Only the scopes returned can be used on
-	 * {@link com.liferay.oauth2.provider.scopes.api.RequiresScope} annotation
+	 * {@link com.liferay.oauth2.apps.api.RequiresScope} annotation
 	 * or in invocations to
-	 * {@link com.liferay.oauth2.provider.scopes.api.ScopeChecker} methods.
+	 * {@link com.liferay.oauth2.apps.api.ScopeChecker} methods.
 	 *
 	 * @param scopeMatcher to filter the application available scopes.
 	 * @return a collection of the available scopes that the implementation
